Updated Import and Examples of Names
Updated the import so that it asks if you want to load in new data. Did this so that once you have things loaded you do not have to keep selecting a folder when you want to run. Also gave examples of what assigning a variable from a structure looks like. Would like to make it so that variables can incorporate names to track which trial and what subject they belong to.